+++ redo.sh
@@ -1,0 +1,3 @@
+pip uninstall xGPR
+rm -rf build *.so xGPR.egg-info
+python setup.py develop