+++ xGPR/random_feature_generation/gpu_rf_gen/cuda_polynomial.pyx
@@ -24,8 +24,6 @@ cdef extern from "convolution_ops/convolution.h" nogil:
 
 
 cdef extern from "poly_ops/polynomial_operations.h" nogil:
-    const char *cudaExactQuadratic_[T](T inArray[], double *outArray, 
-                    int inDim0, int inDim1)
     const char *approxPolynomial_[T](int8_t *radem, T reshapedX[],
         T copyBuffer[], T chiArr[], double *outArray,
         int polydegree, int reshapedDim0, int reshapedDim1,
@@ -248,59 +246,3 @@ def gpuPolyFHT(reshapedX, radem, chiArr, outputArray, int polydegree,
 
     else:
         raise ValueError("Inconsistent array types passed to wrapped C++ function.")
-
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-def cudaExactQuadratic(inputArray, outputArray,
-                int numThreads):
-    """Wraps C++ operations for generating features for an exact
-    quadratic.
-
-    Args:
-        inputArray (ndarray): The input data. This is not modified.
-        outputArray (ndarray): The output array. Must have the appropriate
-            shape such that all of the quadratic polynomial features can
-            be written to it. The last column is assumed to be saved for 1
-            for a y-intercept term.
-        num_threads (int): Number of threads to use for FHT. Not used for gpu,
-            merely kept here for consistency with CPU version.
-
-    Raises:
-        ValueError: A ValueError is raised if unexpected or invalid inputs are supplied.
-    """
-    cdef const char *errCode
-    cdef uintptr_t addr_output = outputArray.data.ptr
-    cdef uintptr_t addr_input = inputArray.data.ptr
-    cdef int numExpectedFeats = int( inputArray.shape[1] * (inputArray.shape[1] - 1) / 2)
-    numExpectedFeats += 2 * inputArray.shape[1] + 1
-
-    if len(inputArray.shape) != 2 or len(outputArray.shape) != 2:
-        raise ValueError("Both inputArray and outputArray for the exact quadratic "
-                "must be 2d arrays.")
-
-    if inputArray.shape[0] == 0:
-        raise ValueError("There must be at least one datapoint.")
-    if inputArray.shape[0] != outputArray.shape[0]:
-        raise ValueError("The number of datapoints in the outputs and the inputs do "
-                "not agree.")
-    if outputArray.shape[1] != numExpectedFeats:
-        raise ValueError("The shape of the output array is incorrect for a quadratic.")
-
-    if not outputArray.flags["C_CONTIGUOUS"] or not inputArray.flags["C_CONTIGUOUS"]:
-        raise ValueError("One or more arguments is not C contiguous.")
-
-    if inputArray.dtype == "float32":
-        errCode = cudaExactQuadratic_[float](<float*>addr_input, <double*>addr_output,
-                        inputArray.shape[0], inputArray.shape[1])
-
-    elif inputArray.dtype == "float64":
-        errCode = cudaExactQuadratic_[double](<double*>addr_input, <double*>addr_output,
-                        inputArray.shape[0], inputArray.shape[1])
-
-    else:
-        raise ValueError("Unexpected types passed to wrapped C++ function.")
-
-    if errCode.decode("UTF-8") != "no_error":
-        raise Exception("Fatal error encountered while performing graph convolution.")