+++ xGPR/random_feature_generation/gpu_rf_gen/poly_ops/polynomial_operations.cu
@@ -1,6 +1,5 @@
 /*
-* Contains functions needed to generate exact quadratic polynomial
-* features and approximate polynomial kernel features on GPU.
+* Contains functions needed to generate approximate polynomial kernel features on GPU.
 */
 
 #include <cuda.h>
@@ -25,9 +24,11 @@ __global__ void polyMultByDiagRademMat(T cArray[], int8_t *rademArray,
     int rVal, position;
     
     position = tid % numElementsPerRow;
-    rVal = rademArray[position];
-    if (tid < numElements)
+
+    if (tid < numElements){
+        rVal = rademArray[position];
         cArray[tid] = cArray[tid] * rVal * normConstant;
+    }
 }
 
 
@@ -44,9 +45,11 @@ __global__ void polyMultAndCopyDiagRademMat(T cArray[], T copyBuffer[],
     int rVal, position;
     
     position = tid % numElementsPerRow;
-    rVal = rademArray[position];
-    if (tid < numElements)
+    
+    if (tid < numElements){
+        rVal = rademArray[position];
         copyBuffer[tid] = cArray[tid] * rVal * normConstant;
+    }
 }
 
 
@@ -66,6 +69,7 @@ __global__ void polyOutArrayCopyTransfer(T copyBuffer[],
     numRowsTraversed = tid / numFreqs;
     numExcess = tid % numFreqs;
     cBuffPosition = numRowsTraversed * numPerRow + numExcess;
+
     if (tid < numElements)
         outArray[tid] = chiArr[numExcess] * copyBuffer[cBuffPosition];
 }
@@ -86,64 +90,13 @@ __global__ void polyOutArrayMultTransfer(T copyBuffer[],
     numExcess = tid % numFreqs;
     cBuffPosition = numRowsTraversed * numPerRow + numExcess;
     chiPosition = repeatNum * numPerRow + numExcess;
+
     if (tid < numElements)
         outArray[tid] *= chiArr[chiPosition] * copyBuffer[cBuffPosition];
 }
 
 
 
-//Generates the features for the exact quadratic.
-template <typename T>
-__global__ void genExactQuadFeatures(T inArray[], double *outArray,
-        int inDim1, int outDim1, int numElements){
-    int pos = blockDim.x * blockIdx.x + threadIdx.x;
-    int rowNum = pos / inDim1;
-    int positionInRow = pos % inDim1;
-    T inVal1 = inArray[pos];
-    T *inPtr = inArray + pos;
-    double *outPtr = outArray + rowNum * outDim1;
-    for (int i=0; i < positionInRow; i++)
-        outPtr += inDim1 + 1 - i;
-    
-    if (pos < numElements){
-        *outPtr = inVal1;
-        outPtr++;
-        for (int i=positionInRow; i < inDim1; i++){
-            *outPtr = inVal1 * *inPtr;
-            outPtr++;
-            inPtr++;
-        }
-    }
-}
-
-
-
-//Performs feature generation for an exact quadratic (i.e. polynomial
-//regression that is exact, not approximated).
-//
-//Note that all of these arrays are already expected to "live" on GPU.
-template <typename T>
-const char *cudaExactQuadratic_(T inArray[], double *outArray,
-                    int inDim0, int inDim1){
-    int numInteractions = (inDim1 * (inDim1 - 1)) / 2;
-    int outDim1 = numInteractions + 1 + 2 * inDim1;
-    int numElements = inDim1 * inDim0;
-    int blocksPerGrid = (numElements + DEFAULT_THREADS_PER_BLOCK - 1) / DEFAULT_THREADS_PER_BLOCK;
-    //cudaProfilerStart();
-
-    //Multiply by D1.
-    genExactQuadFeatures<T><<<blocksPerGrid, DEFAULT_THREADS_PER_BLOCK>>>(inArray, outArray, 
-                                 inDim1, outDim1, numElements);
-    
-
-    //cudaProfilerStop();
-    return "no_error";
-}
-//Instantiate templates explicitly so wrapper can use.
-template const char *cudaExactQuadratic_<float>(float inArray[], double *outArray, 
-                    int inDim0, int inDim1);
-template const char *cudaExactQuadratic_<double>(double inArray[], double *outArray, 
-                    int inDim0, int inDim1);